Backstory is on the front page in the sidebar
Still figuring out vue a bit so had a bit troubles figuring out how to make it pop up on load. However, this temporary fix will have to do for the weekend. I actually like the idea of having a button the player can go back to on the sidebar.